data-sentinel/backend/requirements.txt
fastapi
google-genai
scikit-learn
uvicorn
